Sign_Language_Converter/requirements.txt
Django==5.0.6
opencv-python==4.10.0.82
ultralytics==8.2.31
numpy==1.26.3
djangorestframework==3.15.1
django-cors-headers==4.3.1